libs/stocks/feature-shell/src/lib/stocks/stocks.component.ts
import { Component } from '@angular/core';
import { FormBuilder, FormGroup, Validators } from '@angular/forms';
import { PriceQueryFacade } from '@coding-challenge/stocks/data-access-price-query';

@Component({
  selector: 'coding-challenge-stocks',
  templateUrl: './stocks.component.html',
  styleUrls: ['./stocks.component.css']
})
export class StocksComponent {
  stockPickerForm: FormGroup;
  symbol: string;
  fromDate: string;
  toDate: string;
  todayDate = new Date()

  quotes$ = this.priceQuery.priceQueries$;

  constructor(private fb: FormBuilder, private priceQuery: PriceQueryFacade) {
    this.stockPickerForm = fb.group({
      symbol: [null, Validators.required],
      fromDate: [null, Validators.required],
      toDate: [null, Validators.required]
    });
  }

  fetchQuote() {
    if (this.stockPickerForm.valid) {
      const { symbol, fromDate, toDate } = this.stockPickerForm.value;
      this.priceQuery.fetchQuote(symbol, fromDate, toDate);
    }
  }

  public maxDate(): Date {
    return (this.stockPickerForm.value.toDate && this.stockPickerForm.value.toDate < this.todayDate) ? this.stockPickerForm.value.toDate
    : this.todayDate;
  }
}
